+++ ctt-frontend/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AppComponent } from './app.component';
-import { AuthGuard } from './auth/auth.guard';
+import { AuthGuard } from './guards/auth.guard';
 import { AuthenticationComponent } from './components/authentication/authentication.component';
 import { EditCategoryComponent } from './components/edit-category/edit-category.component';
